Refactor prediction service into layered architecture
