+++ src/oracle/application/__init__.py
@@ -1,0 +1,3 @@
+"""Application layer for Oracle."""
+
+__all__ = []
